requirements.txt
streamlit
sounddevice
scipy
pydub
google-generativeai
SpeechRecognition
pydub
openai
google-generativeai
mutagen
streamlit-webrtc
av
numpy
gtts
librosa
